8Feb2016

+++ DPTS/DPTS.Domain/Entities/Address.cs
@@ -30,9 +30,10 @@ namespace DPTS.Domain.Entities
         public string FaxNumber { get; set; }
 
         public int? Doctor_Id { get; set; }
-         
-        public decimal Latitude { get; set; }
-        public decimal Longitude { get; set; }
+
+        public double Latitude { get; set; }
+
+        public double Longitude { get; set; }
 
         public virtual ICollection<AddressMapping> AddressMappings { get; set; }
